Implement top level caching for palomabot.
Fixes #123
